fix: refactor developmentcss file. split to several files
