+++ imports/startup/server/firstRunCss.js
@@ -1,4 +1,7 @@
-import { leftSideBarCss, systemcss, usercss, menuLinksCss } from "../../../server/developmentcss";
+import leftSideBarCss from "../../../server/defaultStyles/leftSideBarCss";
+import menuLinksCss from "../../../server/defaultStyles/menuLinksCss";
+import systemCss from "../../../server/defaultStyles/systemCss";
+import userCss from "../../../server/defaultStyles/userCss";
 import mongoCss from "../../collections/css";
 import { Meteor } from "meteor/meteor";
 
@@ -7,9 +10,12 @@ export default function firstRunCSS() {
     return;
   }
 
-  if (!mongoCss.find().count()) {
-    mongoCss.insert(systemcss);
-    mongoCss.insert(usercss);
+  if (!mongoCss.find({ type: "system" }).count()) {
+    mongoCss.insert(systemCss);
+  }
+
+  if (!mongoCss.find({ type: "board" }).count()) {
+    mongoCss.insert(userCss);
   }
 
   if (!mongoCss.find({ type: "leftSideBar" }).count()) {
