+++ server/defaultStyles/menuLinksCss.js
@@ -1,0 +1,29 @@
+const menuLinksCss = {
+  type: "menuLinks",
+  menuLinks: {
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "space-between",
+    flexGrow: 1,
+    marginTop: "46px"
+  },
+  menuLinkItem: {
+    display: "flex",
+    marginBottom: "14px",
+    width: "100%"
+  },
+  menuItemText: {
+    width: "100%"
+  },
+  acitve: {
+    backgroundColor: "#2a9bdc"
+  },
+  rowStyle: {
+    paddingLeft: 0,
+    listStyle: "none",
+    marginTop: 0,
+    marginBottom: "1em"
+  }
+};
+
+export default menuLinksCss;
